% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pjenyuh-package.R
\docType{package}
\name{hw04pjenyuh-package}
\alias{hw04pjenyuh}
\alias{hw04pjenyuh-package}
\title{hw04pjenyuh: Homework 04 for STAT-613}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/STAT-413-613-21S/hw04pjenyuh}
  \item Report bugs at \url{https://github.com/STAT-413-613-21S/hw04pjenyuh/issues}
}

}
\author{
\strong{Maintainer}: Jen-Yu Huang \email{jh6669a@student.american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
