% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_n.R
\name{myseq_n}
\alias{myseq_n}
\title{Get the nth element of a certain recursive sequence}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{a vector containing the first three numeric elements of this sequence}

\item{n}{a positive integer}
}
\value{
element n
}
\description{
Element n is the sum of element n-1 and the difference between elements n-2 and n-3 divided by n
}
\examples{
myseq_n(x = c(2, 4, 3), n = 7)
}
