% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_plot.R
\name{myseq_plot}
\alias{myseq_plot}
\title{Plot the output value of 'myseq_n'}
\usage{
myseq_plot(input_df)
}
\arguments{
\item{input_df}{a data frame with 4 columns}
}
\value{
plot
}
\description{
Plot the output value of 'myseq_n'
}
\examples{
myseq_plot(data.frame(x = 2, y = 4, z = 3, n = 3:9)), myseq_plot(data.frame(2, 4, 3, 3:9))
}
