% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pjenyuh-package.R
\docType{package}
\name{hw04pjenyuh-package}
\alias{hw04pjenyuh}
\alias{hw04pjenyuh-package}
\title{hw04pjenyuh: Homework 04 for STAT-613}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jen-Yu Huang \email{jh6669a@student.american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
